docs(progressbar): categorize under Controls
