--- src/Progress/Progress.mdx
@@ -1,49 +1,0 @@
----
-name: Progress
-menu: Components
----
-
-import { Progress } from './Progress';
-
-# Progress
-
-## Usage
-
-#### Default
-
-<Playground>
-  {() => {
-    {
-      /*
-      @toDo
-      for some reason the value isn't being updated, check this later
-    */
-    }
-    const [percent, setPercent] = React.useState(0);
-    React.useEffect(() => {
-      const timer = setInterval(() => {
-        setPercent(previousPercent => {
-          if (previousPercent === 100) {
-            return 0;
-          }
-          const diff = Math.random() * 10;
-          return Math.min(previousPercent + diff, 100);
-        });
-      }, 1000);
-      return () => clearInterval(timer);
-    }, []);
-    return <Progress variant='default' value={Math.floor(percent)} />;
-  }}
-</Playground>
-
-## API
-
-### Import
-
-```
-import { Progress } from 'react95'
-```
-
-### Props
-
-<Props of={Progress} />
