+++ src/index.ts
@@ -19,7 +19,7 @@ export * from './MenuList/MenuList';
 export * from './Monitor/Monitor';
 export * from './NumberField/NumberField';
 export * from './Panel/Panel';
-export * from './Progress/Progress';
+export * from './ProgressBar/ProgressBar';
 export * from './Radio/Radio';
 export * from './ScrollView/ScrollView';
 export * from './Select/Select';
@@ -50,3 +50,4 @@ export * from './legacy/Divider';
 export * from './legacy/Fieldset';
 export * from './legacy/List';
 export * from './legacy/ListItem';
+export * from './legacy/Progress';
