+++ src/legacy/Progress.tsx
@@ -1,0 +1,7 @@
+import { ProgressBar, ProgressBarProps } from '../ProgressBar/ProgressBar';
+
+/** @deprecated Use `ProgressBarProps` */
+export type ProgressProps = ProgressBarProps;
+
+/** @deprecated Use `ProgressBar` */
+export const Progress = ProgressBar;
